# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Johny Tuna",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/cart/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Получение информации о корзине",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/cart/update": {
            "post": {
                "description": "Использовать этот запрос в случае обнавления товара в корзине, для добавления нового товара или удаления товара из корзины (count = 0)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Получение информации о корзине",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "обновление корзины",
                        "name": "update_cart",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateCart"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/order/place": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "Получение информации о корзине",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token}",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/products/getByCategory": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получение продуктов по категории",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id категории",
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/search/products": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Получение продуктов из поиска",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя продукта (поиск)",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "dto.UpdateCart": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}